requirements.txt
asgiref==3.9.1
dj-database-url==2.3.0
Django==4.2.10
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-taggit==5.0.1
django-widget-tweaks==1.5.0
gunicorn==21.2.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.9.0
